contacts retrieved and rendered
