+++ App/navigators/ScanStack.tsx
@@ -2,14 +2,14 @@ import React from 'react'
 import { createStackNavigator } from '@react-navigation/stack'
 
 import defaultStackOptions from './defaultStackOptions'
-import Workflow from '../screens/Workflow'
+import Scan from '../screens/Scan'
 
 const Stack = createStackNavigator()
 
 function ScanStack() {
   return (
     <Stack.Navigator screenOptions={defaultStackOptions}>
-      <Stack.Screen name="Scan" component={Workflow} />
+      <Stack.Screen name="Scan" component={Scan} />
     </Stack.Navigator>
   )
 }
